requirements.txt
certifi==2017.11.5
chardet==3.0.4
click==6.7
Flask==0.12.2
future==0.16.0
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
line-bot-sdk==1.5.0
MarkupSafe==1.0
requests==2.18.4
urllib3==1.22
Werkzeug==0.12.2
